+++ docs/multilogin-api/spec/code-samples/csharp/HttpClient/block_07.cs
@@ -1,0 +1,14 @@
+{
+  "data": {
+    "browser_type": "mimic",
+    "core_version": 132,
+    "id": "d04aa438-d887-11ef-8d6d-0a0027000012",
+    "is_quick": true,
+    "port": "55579"
+  },
+  "status": {
+    "error_code": "",
+    "http_code": 200,
+    "message": "Quick profile started successfully"
+  }
+}