+++ docs/multilogin-api/spec/code-samples/java/OkHttp/block_04.java
@@ -1,0 +1,28 @@
+{
+    "browser_type": "<string>",
+    "os_type": "<string>",
+    "automation": "<string>",
+    "auto_update_core": <boolean>,
+    "core_version": <number>,
+    "is_headless": <boolean>,
+    "parameters": {
+        "flags": {
+            "audio_masking": "value",
+            "fonts_masking": "value",
+            "geolocation_masking": "value",
+            "geolocation_popup": "value",
+            "graphics_masking": "value",
+            "graphics_noise": "value",
+            "localization_masking": "value",
+            "media_devices_masking": "value",
+            "navigator_masking": "value",
+            "ports_masking": "value",
+            "proxy_masking": "value",
+            "screen_masking": "value",
+            "timezone_masking": "value",
+            "webrtc_masking": "value",
+            "canvas_noise" :"value",
+            "startup_behavior" :"value"
+        }
+        "fingerprint": {}
+}