+++ docs/multilogin-api/spec/code-samples/python/http_client/block_03.py
@@ -1,0 +1,14 @@
+{
+  "data": {
+    "browser_type": "mimic",
+    "core_version": 132,
+    "id": "81b5627a-1212-4016-9467-3dbe4d6f78eb",
+    "is_quick": false,
+    "port": "55513"
+  },
+  "status": {
+    "error_code": "",
+    "http_code": 200,
+    "message": "Profile started successfully"
+  }
+}